Added constructor for CharacterEntity
Added constructor for CharacterEntity and fixed an integer division bug.

+++ PantheonPrototype/PantheonPrototype/HUD.cs
@@ -65,7 +65,7 @@ namespace PantheonPrototype
             // Set the width of the Armor Bar with respect to the current percent of the player's armor. (Player not implemented yet)
             try
             {
-                hudItems[0].Coordinates = new Rectangle(hudItems[0].Coordinates.X, hudItems[0].Coordinates.Y, (int)(hudItems[0].DefaultWidth * (player.CurrentArmor / player.TotalArmor)), hudItems[0].Coordinates.Height);
+                hudItems[0].Coordinates = new Rectangle(hudItems[0].Coordinates.X, hudItems[0].Coordinates.Y, (int)(hudItems[0].DefaultWidth * ((float)player.CurrentArmor / player.TotalArmor)), hudItems[0].Coordinates.Height);
             }
             catch (DivideByZeroException)
             {
